requirements.txt
beautifulsoup4>=4.13.3,<5.0.0
blinker>=1.9.0,<2.0.0
click>=8.1.8,<8.2.0
coverage>=7.7.1,<7.8.0
dotenv>=0.9.9,<1.0.0
Flask>=3.1.0,<3.2.0
Flask-Login>=0.6.3,<1.0.0
Flask-Session>=0.8.0,<1.0.0
flask-talisman>=1.1.0,<2.0.0
gunicorn>=23.0.0,<24.0.0
iniconfig>=2.1.0,<2.2.0
itsdangerous>=2.2.0,<2.3.0
Jinja2>=3.1.6,<3.2.0
MarkupSafe>=3.0.2,<3.1.0
mmh3>=5.1.0,<6.0.0
packaging>=24.2,<24.3
pluggy>=1.5.0,<1.6.0
psycopg2-binary>=2.9.10,<3.0.0
pytest>=8.3.5,<8.4.0
redis>=5.2.1,<6.0.0
requests>=2.32.3,<3.0.0
urllib3>=2.4.0,<3.0.0
Werkzeug>=3.1.3,<3.2.0
pyyaml>=6.0.1
flask-limiter>=3.12
